OAK-9687 (refactor): move more generic stuff out of the query package
